adding graph and some minor UI changes

+++ app/src/main/java/com/losdol/dashboardonlysonik/deviceData.java
@@ -6,13 +6,14 @@ import java.util.HashMap;
 
 public class deviceData {
     public Boolean online;
-    public HashMap<String, Long> sensors;
+    public HashMap<String, Float> sensors;
     public String username;
     public Timestamp timestamp;
 
+
     public deviceData() {}
 
-    public deviceData(Boolean online, HashMap<String, Long> sensors, String username, Timestamp timestamp){
+    public deviceData(Boolean online, HashMap<String, Float> sensors, String username, Timestamp timestamp){
         this.online = online;
         this.sensors = sensors;
         this.timestamp = timestamp;
@@ -23,7 +24,7 @@ public class deviceData {
         return username;
     }
 
-    public HashMap<String, Long> getSensors() {
+    public HashMap<String, Float> getSensors() {
         return sensors;
     }
